requirements.txt
pandas
numpy
beautifulsoup4
requests
sentence-transformers
faiss-cpu
flask
streamlit
scikit-learn
openpyxl
